+++ lib/govuk_publishing_components/presenters/machine_readable/article_schema.rb
@@ -29,7 +29,7 @@ module GovukPublishingComponents
               "url" => page.logo_url,
             }
           }
-        }.merge(image_schema).merge(author_schema).merge(body)
+        }.merge(image_schema).merge(author_schema).merge(body).merge(is_part_of)
       end
 
     private
@@ -66,6 +66,37 @@ module GovukPublishingComponents
 
       def publishing_organisation
         page.content_item.dig("links", "primary_publishing_organisation").to_a.first
+      end
+
+      def is_part_of
+        return {} unless linked_step_by_steps.any?
+        {
+          "isPartOf" => step_by_step_schemas
+        }
+      end
+
+      def linked_step_by_steps
+        # We could include `related_to_step_navs` eventually too, but initially
+        # link to those that we render in the "step_by_step_nav_related" component
+        @step_by_steps ||= page.content_item.dig("links", "part_of_step_navs").to_a
+      end
+
+      def step_by_step_schemas
+        linked_step_by_steps.map do |step_by_step|
+          step_by_step_page = linked_page(step_by_step)
+          structured_data = HowToSchema.new(step_by_step_page.canonical_url).structured_data
+
+          structured_data.merge(image_schema)
+        end
+      end
+
+      def linked_page(step_by_step)
+        Page.new(
+          content_item: step_by_step,
+          schema: :article,
+          logo_url: page.logo_url,
+          image_placeholders: page.image_placeholders
+        )
       end
     end
   end